temp_company_research/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
trafilatura==1.12.2
openai==1.46.0
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
